Reworks button listener event
The numeric button currently print to the screen. The buttons that
correspond to the four basic operations are passed onto an independent
variable stored in the Calculator class. Additionally, if the user
presses the same operator again, it causes the variable to reset to
an empty string so that the user can continue on working on the value.

An accompanying boolean variable called "activeOperator" was also
created in the Calculator class, which will be used to trigger the
processing of the operation, but it might be further simplified so as
not to need it.

The calculator cannot yet process any operations.