Introduces JS scripts
A basic set of 3 js scripts was added, grouped in a specific folder.
The first script, which is directly linked to the html file, is called
"main.js", which only purpose is to initiate the class set in another
js module.
The second module hosts the class initiated in main.js, namely MyUI. Its
purpose is to add functionality to the basic UI on the HTML, such as:
adding event listeners to the buttons and refreshing the screen, etc.
So far the script does not support processing any operation and the most
that it can do is pass the values of the button numbers to a variable
inside the Calculator class in the third module. It also loads the
starting values for the screen, upon first loading the page.
The third script hosts the Calculator class, which is used by the module
ui.js. The class contains functions related to the
calculator itself, namely: addition, subtraction, and so on, as well as
the variables that contains the data needed for both carrying out
operations and for displaying values to the screen. The last of which
are fetched by functions within the MyUI class.